requirements.txt
aioredis==1.3.1
allure-pytest==2.9.45
allure-python-commons==2.9.45
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.7
django-celery-beat==2.2.1
django-cors-headers==3.9.0
django-filter==21.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.8.1
incremental==21.3.0
iniconfig==1.1.1
kombu==5.1.0
msgpack==1.0.2
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.21
psycopg2==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.4.0
pytest-forked==1.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zipp==3.6.0
zope.interface==5.4.0
